fix(docker): force MariaDB to use utf8mb4 by default
This is a quite simple fix. When creating the container with
docker-compose just issue a command to change the server charset to
utf8mb4 and collation ot utf8mb4_unicode_ci. This should suffice to
accommodate all combinations of latin and various characters that have
found their way into domain names.

Closes #8