fix(model:PiosResult): force short date format
For some reason CultureInfo is braindead, and does not like it when your
system's regional settings have ISO format for short date, this just
forces the date format used by the .gr Registry. This is unlikely to
change in the near future, so I also added a note to keep this in mind.