fix(svc): weird duplicate entity exception
When saving everything related to a fresh Snapshot in one go, there was
some weird error for a duplicate IP address, which wasn't actually
present in the database. The remedy is to call SaveChanges() for each
individual entity. (Need to check how this impacts performance in the
future, but for now it works.)

Closes #21